webproject: Add gunicorn hook for stderr logging
By default when gunicorn is invoked with --daemon, it redirects 0, 1, 2
file descriptors to /dev/null and as a result the logs printed on
stdout/stderr are lost. This commit adds a gunicorn hook that uses the
already opened file of the first registered logging FileHandler, i.e.
the file from the '--log-file' Gunicorn configuration option.
To enable this hook use
--config=/etc/synnefo/gunicorn-hooks/gunicorn-stderr-logging.py option.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>