pithos: Delete mapfiles when purging nodes
Versions in database have a Merkle hash of a given content
address-able object and a backend-specific mapfile that indexes the
corresponding block hashes that the object consists of. When purging
a node, deleting a container, or explicitly removing a version we
call the corresponding store method (i.e. map_delete()) for
deleting the version's mapfile.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>